// src/types/productType.ts
export interface ProductType {
    id:string
    title: string
    price: number
    thumbnail: string
    description: string
    stock: number
}

export interface AddProductType {
    // id: number
    title: string
    price: number
    thumbnail: string
    description: string
    stock: number

}